Add Grad widget

+++ src/enums/fg.rs
@@ -19,7 +19,7 @@ pub enum Fg {
     Magenta,
     Cyan,
     White,
-    RGB(usize, usize, usize),
+    RGB(u8, u8, u8),
     Default,
 }
